requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
pillow
lime
scikit-image

